Time-Series-Forecasting-main/Deployment/Docker/requirements.txt
django
djangorestframework
joblib
pandas
scikit-learn